+++ package.json
@@ -24,6 +24,7 @@
 		"react": "17.0.2",
 		"react-dom": "17.0.2",
 		"react-hook-form": "^7.33.0",
+		"react-loading": "^2.0.3",
 		"styled-components": "^5.3.3"
 	},
 	"devDependencies": {
