+++ src/types/authProviderType.ts
@@ -4,6 +4,7 @@ import { ReactNode } from 'react';
 export type AuthContextProps = {
 	currentUser: User | null | undefined;
 	isAuthenticated: boolean;
+	isLoading: boolean;
 };
 
 export type AuthProviderProps = {
